third_party/WebKit/LayoutTests/fast/multicol/vertical-rl/offset-top-and-left-at-boundaries-nested.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div style="position:relative; writing-mode:vertical-rl; columns:3; column-fill:auto; column-gap:10px; width:100px; height:350px;">
    <div style="columns:2; column-gap:10px;">
        <div id="first" style="width:100px;"></div>
        <div id="second" style="width:100px;"></div>
        <div id="third" style="width:100px;"></div>
        <div id="fourth" style="width:100px;"></div>
    </div>
    <div style="columns:2; column-gap:10px;">
        <div id="fifth" style="width:100px;"></div>
        <div id="sixth" style="width:100px;"></div>
    </div>
</div>

<script>
    test(function() {
        assert_equals(document.getElementById('first').offsetLeft, 0);
        assert_equals(document.getElementById('first').offsetTop, 0);
    }, "offsetLeft and offsetTop in vertical-rl at nested column boundaries, first element");

    test(function() {
        assert_equals(document.getElementById('second').offsetLeft, 0);
        assert_equals(document.getElementById('second').offsetTop, 60);
    }, "offsetLeft and offsetTop in vertical-rl at nested column boundaries, second element");

    test(function() {
        assert_equals(document.getElementById('third').offsetLeft, 0);
        assert_equals(document.getElementById('third').offsetTop, 120);
    }, "offsetLeft and offsetTop in vertical-rl at nested column boundaries, third element");

    test(function() {
        assert_equals(document.getElementById('fourth').offsetLeft, 0);
        assert_equals(document.getElementById('fourth').offsetTop, 180);
    }, "offsetLeft and offsetTop in vertical-rl at nested column boundaries, fourth element");

    test(function() {
        assert_equals(document.getElementById('fifth').offsetLeft, 0);
        assert_equals(document.getElementById('fifth').offsetTop, 240);
    }, "offsetLeft and offsetTop in vertical-rl at nested column boundaries, fifth element");

    test(function() {
        assert_equals(document.getElementById('sixth').offsetLeft, 0);
        assert_equals(document.getElementById('sixth').offsetTop, 300);
    }, "offsetLeft and offsetTop in vertical-rl at nested column boundaries, sixth element");
</script>
